# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Storages API",
        "description": "Warehouse management system API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List all products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "sale_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ahmad Earl Grey 500g"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "example": 45000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or name not unique"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ahmad Earl Grey 500g"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "example": 45000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/products/remaining": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Remaining product quantities as of a given date",
                "operationId": "87236ec30997c24b85a954035bedd0b3",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remaining quantities per product at the given date",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Ahmad Earl Grey 500g"
                                            },
                                            "category_name": {
                                                "type": "string",
                                                "example": "Black Tea"
                                            },
                                            "sale_price": {
                                                "type": "number",
                                                "example": 45000
                                            },
                                            "remaining_qty": {
                                                "type": "integer",
                                                "example": 80
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ahmad Earl Grey 500g"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "sale_price": {
                        "type": "number",
                        "example": 45000
                    }
                },
                "type": "object"
            }
        }
    },
...
}